implement push message handling

+++ usecase/igateway/iline_repositry.go
@@ -7,5 +7,5 @@ import (
 
 type ILineRepository interface {
 	FindUserById(id string, botType dto.BotType) (*entity.User, error)
-	GetQuotaComsuption(botType dto.BotType) (int, error)
+	GetQuotaComsuption(botType dto.BotType) (int64, error)
 }
